Add readFile tests
